SF, remove old and add missing outputs
