+++ outputs/elixir_10_1000000000
@@ -1,0 +1,10 @@
+1:0.100010044
+2:0.100007356
+3:0.09998641
+4:0.10000192
+5:0.099991963
+6:0.100013923
+7:0.099998666
+8:0.100000696
+9:0.099992962
+10:0.09999606
