+++ outputs/kotlin_10_1000000000
@@ -1,0 +1,10 @@
+5:0.100008517
+8:0.10000897
+3:0.100002436
+2:0.100014119
+4:0.099997343
+6:0.099989069
+10:0.099990483
+1:0.100006535
+7:0.099983049
+9:0.099999479
